create lr files
