+++ src/parser/Makefile
@@ -7,6 +7,7 @@ all: sol_dfa.o sol_pattern.o sol_ll1.o
 sol_dfa.o: sol_dfa.c sol_common.h sol_hash.o sol_set.o
 sol_pattern.o: sol_pattern.c sol_dfa.o sol_list.o
 sol_ll1.o: sol_ll1.c sol_common.h sol_hash.o sol_list.o sol_stack.o sol_rbtree.o sol_rbtree_iter.o
+sol_slr.o: sol_slr.c sol_common.h sol_hash.o sol_list.o sol_stack.o sol_rbtree.o sol_rbtree_iter.o sol_rb_tuple.o
 
 test_dfa: test_dfa.c sol_dfa.o sol_common.h sol_hash.o sol_set.o sol_utils.o  Hash_fnv.c Hash_murmur.c
 	if [ ! -d output ]; then mkdir output; fi
@@ -19,6 +20,8 @@ test_pattern: test_pattern.c sol_pattern.o sol_dfa.o sol_hash.o sol_set.o sol_ut
 test_ll1: test_ll1.c sol_ll1.o sol_stack.o sol_list.o sol_hash.o sol_dl_list.o sol_rbtree.o sol_rbtree_iter.o
 	if [ ! -d output ]; then mkdir output; fi
 	$(CC) $(CFLAGS) -o output/$@ $^
+
+test_slr: test_slr.c sol_slr.o
 
 .PHONY: clean
 clean:
