+++ src/parser/test_slr.c
@@ -1,0 +1,9 @@
+#include <stdio.h>
+#include "sol_slr.h"
+
+int main()
+{
+    SolSLRParser *p = solSLRParser_new();
+    solSLRParser_free(p);
+    return 0;
+}
